pyproject.toml
[tool.poetry]
name = "social_network"
version = "0.1.0"
description = ""
authors = ["dmitriy <dmitriy@bessalov.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
Pillow = "^8.0.1"
django-ckeditor = "^6.0.0"
djangorestframework = "^3.12.2"
djoser = "^2.1.0"
django-cors-headers = "^3.7.0"
django-rest-swagger = "^2.2.0"
python-dotenv = "^0.17.0"
psycopg2-binary = "^2.8.6"
daphne = "^3.0.1"
django-ninja = "^0.12.2"
orjson = "^3.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
